Make composer WordPress installer robust enough to delete all dotfiles
Closes: #121

+++ scripts/WordPress/Installer.php
@@ -6,7 +6,8 @@ use Composer\Script\Event;
 
 class Installer {
   /**
-   * Remove wp-content from WordPress and symlink it to correct location
+   * Fully delete the default wp-content from WordPress and replace it with a
+   * symlink it to ../wp-content. This also removes the default theme and plugins.
    *
    * @param Composer\Script\Event $event - This is the way composer talks to it's plugins
    */
@@ -37,8 +38,16 @@ class Installer {
    * @param String $dir - path to folder to be destroyed
    */
   public static function rrmdir($dir) {
-    foreach (glob($dir . '/.htaccess') as $file) {
-      unlink($file);
+    // Delete all dotfiles first since the latter '*' will not catch them
+    // this is needed for exampel to delete:
+    //   - plugins/akismet/.htaccess
+    //   - themes/twentytwenty/.stylelintrc.json
+    foreach (glob($dir . '/.*') as $file) {
+      // The glob above will also match directories, so check match type first
+      // and only delete files
+      if (is_file($file)) {
+        unlink($file);
+      }
     }
     foreach (glob($dir . '/*') as $file) {
       if (is_dir($file)) {
